backend/requirements.txt
fastapi
pandas
uvicorn
openpyxl
python-dotenvpython-dotenv
groq
scikit-learn